+++ NEWS.md
@@ -2,3 +2,5 @@ In upcoming version
 
 
   * Updated the limits of the residualplot function to ensure that the blurred area are always within the plot region
+  
+  * Updated the residualplot function to accommodate glm objects with binomial and poisson families